feat: calculate item percentcod on checkout

+++ src/Actions/OrderCreateAction.php
@@ -356,7 +356,7 @@ namespace Cdek\Actions {
 
             if ($shouldPay !== null) {
                 if ($shouldPay !== 0) {
-                    $payment['value'] = (int)(($shouldPay / 100) * $cost);
+                    $payment['value'] = (float)(($shouldPay / 100) * $cost);
 
                     if($product->is_taxable()){
                         $taxCost = $shouldConvert === null ? (float)$item->get_total_tax() :
